Removed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/physics@10242 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/karts/kart.cpp
@@ -755,7 +755,7 @@ float Kart::getActualWheelForce()
  */
 bool Kart::isOnGround() const
 {
-    return (m_vehicle->getNumWheelsOnGround() == m_vehicle->getNumWheels()
+    return ((int)m_vehicle->getNumWheelsOnGround() == m_vehicle->getNumWheels()
           && !playingEmergencyAnimation());
 }   // isOnGround
